Add dynamic ResourceItemDescriptors and measured value items
The PR allows creation of ResourceItemsDescriptors from `generic/items` JSON files.
Works the same as the items defined in C++ code.

Dynamic items are only added if they aren't already defined. Later on we can remove items defined in C++ code, but need to reroute const char* suffix pointers like `RStateOn`.
The PR also defines the items described in https://github.com/dresden-elektronik/deconz-rest-plugin/pull/6673#issuecomment-1521865951

Update measured_value items according to discussion


Remove measure_value items from this PR

+++ utils/utils.h
@@ -57,7 +57,7 @@ struct RestData
     int integer;
     uint uinteger;
     QString string;
-    float real;
+    double real;
     bool valid = false;
 };
